Allow usage without extending String
