+++ lib/whatlanguage.rb
@@ -99,13 +99,3 @@ class WhatLanguage
     define_method(:to_lowercase) { |str| UnicodeUtils.casefold(str) }
   end
 end
-
-class String
-  def language
-    WhatLanguage.new(:all).language(self)
-  end
-
-  def language_iso
-    WhatLanguage.new(:all).language_iso(self)
-  end
-end
