+++ lib/whatlanguage/string.rb
@@ -1,0 +1,11 @@
+require 'whatlanguage'
+
+class String
+  def language
+    WhatLanguage.new(:all).language(self)
+  end
+
+  def language_iso
+    WhatLanguage.new(:all).language_iso(self)
+  end
+end
